test: fix error in frontend unit test
